Work on abstracting logic
